+++ src/App.tsx
@@ -1,7 +1,7 @@
-import "./App.css";
-import { Card } from "./components/Card";
-import Header from "./components/Header";
-import DottedLine from "./assets/dotted-line.svg";
+import './App.css';
+import { Card } from './components/Card';
+import Header from './components/Header1';
+import DottedLine from './assets/dotted-line.svg';
 
 function App() {
   return (
@@ -9,11 +9,7 @@ function App() {
       <div className="h-screen w-screen bg-white bg-gradient-to-t from-[#15A09166] via-[#15A09166] to-white flex justify-center items-center flex-col">
         <Header></Header>
         <Card />
-        <img
-          className="fixed left-4 bottom-0"
-          src={DottedLine}
-          alt="Dotted Line"
-        />
+        <img className="fixed left-4 bottom-0" src={DottedLine} alt="Dotted Line" />
       </div>
     </>
   );
