+++ src/pages/Home.tsx
@@ -1,4 +1,4 @@
-import Header from "../components/Header";
+import Header from '../components/Header1';
 
 const Home = () => {
   return (
